requirements.txt
fastapi==0.115.12
google_api_python_client==2.169.0
google-adk==0.5.0
protobuf>=4.25.0,<5.0.0
python-dotenv==1.1.0
Requests==2.32.3
vertexai==1.43.0
google-cloud-aiplatform[agent_engines, adk]==1.91.0
slack-sdk==3.29.0
google-cloud-logging==3.10.0
pytest==8.1.1
pytest-asyncio==0.23.6
deprecated>=1.2.14
cloudpickle==3.1.1
pydantic==2.11.7